DataLoggerGUI_2.0/online_access_tool/src/test/java/nl/han/oose/project/crosscutting_concerns/AuthenticationTest.java
package nl.han.oose.project.crosscutting_concerns;
import nl.han.oose.project.business.exceptions.UnauthorizedException;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import static org.junit.jupiter.api.Assertions.*;
import static org.junit.jupiter.api.Assertions.assertEquals;

public class AuthenticationTest {
     private Authentication sut;
     private String password = "testPassword";

     @BeforeEach
     public void setup(){
          this.sut = new Authentication();
     }

     @Test
     public void hashPassword_ReturnsHashedPassword(){
          // Act
          String hashedPw = this.sut.hashPassword(password);

          // Assert
          assertNotEquals(password, hashedPw);
     }

     @Test
     public void checkPassword_ThrowsUnauthorized_WhenPasswordDoesntMatch(){
          // Arrange
          String notTheSamePassw = "password";
          String errorMsg = "The password you entered is incorrect.";

          // Act
          Exception exception = assertThrows(UnauthorizedException.class, () -> {
               this.sut.checkPassword(password, notTheSamePassw);
          });

          // Assert
          assertEquals(UnauthorizedException.class, exception.getClass());
          assertEquals(errorMsg, exception.getMessage());
     }
}